Fixed ClearDepth issue on ATI.  Honestly, it should have also been an issue on NVIDIA since glClearDepthf is only in GL 4.1.

+++ Source/Renderer/GL3x/ContextGL3x.cs
@@ -161,7 +161,7 @@ namespace OpenGlobe.Renderer.GL3x
 
             if (_clearDepth != clearState.Depth)
             {
-                GL.ClearDepth(clearState.Depth);
+                GL.ClearDepth((double)clearState.Depth);
                 _clearDepth = clearState.Depth;
             }
